requirements.txt
asgiref==3.10.0
babel==2.17.0
certifi==2025.11.12
charset-normalizer==3.4.4
Django==5.2.8
djangorestframework==3.16.1
idna==3.11
pillow==12.0.0
psycopg2-binary==2.9.11
requests==2.32.5
sqlparse==0.5.3
tkcalendar==1.6.1
tzdata==2025.2
urllib3==2.5.0
gunicorn==23.0.0
django-cors-headers==4.5.0
python-decouple==3.7
whitenoise==6.5.0
dj-database-url==1.1.0

